# model/normalizer: Enforce valid utf-8 during normalization.