An attempt at enforcing valid utf8 strings for span contents after normalization.

It works but has a non-negligible impact which I couldn't avoid by writing the same code with several variations (my own loop, etc...). Most of the time and allocations are actually inside the Rune handling code of the stdlib it seems.

Impact on my MBP:

Impact with no invalid utf-8 strings:
```
$ benchstat old.txt new-valid.txt
name             old time/op    new time/op    delta
Normalization-4     549ns ± 8%     671ns ± 0%  +22.13%  (p=0.016 n=5+4)

name             old alloc/op   new alloc/op   delta
Normalization-4      792B ± 0%      792B ± 0%     ~     (all equal)

name             old allocs/op  new allocs/op  delta
Normalization-4      9.00 ± 0%      9.00 ± 0%     ~     (all equal)
```

Impact with invalid utf-8 strings in Resource, Type and Meta:
```
$ benchstat old.txt new-invalid.txt
name             old time/op    new time/op    delta
Normalization-4     549ns ± 8%    1344ns ± 0%  +144.68%  (p=0.008 n=5+5)

name             old alloc/op   new alloc/op   delta
Normalization-4      792B ± 0%     1296B ± 0%   +63.64%  (p=0.008 n=5+5)

name             old allocs/op  new allocs/op  delta
Normalization-4      9.00 ± 0%     17.00 ± 0%   +88.89%  (p=0.008 n=5+5)
```