@AlexJF 2018-06-05
Done, didn't know about this benchstat thing, useful :)