@AlexJF 2018-06-06
Oh `strings.Builder` apparently requires go 1.10. Do we need compatibility with 1.9 or can we update the CI?