@gbbr 2018-06-06
It's definitely not a problem to update CI. AFAIK we are aiming to have the same version in CI as we do in [Agent 6](https://github.com/DataDog/datadog-agent).

@derekwbrown or @masci can you guys please confirm that it's ok for us to have a requirement for go1.10? (tldr; to use `strings.Builder`)