@masci 2018-06-06
Ping @olivielpeau for review