@AlexJF 2018-06-06
My 2 cents: 

We can't really control what gets sent to us and given that traces are encoded via protobuf and this was merged (https://github.com/golang/protobuf/pull/499), then any non-utf8 traces would be immediately dropped at the agent during serialization time when/if we upgrade to latest versions.

This might be acceptable behaviour or not, that's more of a product decision. I'll just comment that, from a user's standpoint, it seems much more friendly to let the traces go through but clearly mark strings with invalid utf8 with `�`. But as this PR shows, this comes with a certain cost.

Another thing we could do is have this utf8 normalization happen later on in the pipeline, after sampling. That could reduce the performance impact but would mean we'll have traces flying through our pipeline with invalid utf8. This might not be a problem currently but if we introduce more parts later on it's something to keep in mind.